# Speed up `deleted_reason` method