@apainintheneck 2022-11-12
I thought I'd add my git version too since that could also be a factor in the speed of the `git log` command.

```sh
$ git version
git version 2.24.3 (Apple Git-128)
```