@apainintheneck 2022-11-13
Okay, I reran the benchmarks with my work laptop and there still seems to be a significant benefit. The big difference seems to be that the newer version of git on the M1 Macbook runs the `git log` command about twice as fast as on my old computer.

```sh
$ pwd
/opt/homebrew/Library/Taps/homebrew/homebrew-core/o/h/L/T/h/homebrew-core
$ time git log --since='1 month ago'\
    --diff-filter=D --max-count=1\
    --format=%H\\\\n%h\\\\n%B -- Formula/unknown_formula.rb

________________________________________________________
Executed in    3.74 secs    fish           external
   usr time    3.65 secs    0.33 millis    3.65 secs
   sys time    0.05 secs    2.31 millis    0.04 secs
```

It seems like this change would still save a few seconds when using the `brew info` and `brew search` commands.

|cmd|pc|before|after|
|--|--|--|--|
|`brew info jjrnl`|old iMac|6.10 secs|2.38 secs|
|`brew info jjrnl`|M1 Macbook|2.73 secs| 894.83 millis|
|`brew info unknown_formula`|old iMac|9.68 secs|2.39 secs|
|`brew info unknown_formula`|M1 Macbook|4.46 secs|2.56 secs|
|`brew search jjrnl`|old iMac|9.80 secs|6.18 secs|
|`brew search jjrnl`|M1 Macbook|7.91 secs|2.06 secs|
|`brew search unknown_formula`|old iMac|15.30 secs|6.20 secs|
|`brew search unknown_formula`|M1 Macbook|6.29 secs| 2.95 secs|

Legend:
<details>
<summary>Old iMac: Brew Config</summary>

```
HOMEBREW_VERSION: 3.6.10-9-ge1b2f68
ORIGIN: https://github.com/Homebrew/brew
HEAD: e1b2f68a6d0c151254f2686b50ec661b973f83b6
Last commit: 2 days ago
Core tap ORIGIN: https://github.com/Homebrew/homebrew-core
Core tap HEAD: 24255ed4817a3b54f91a1aa893140c8017a67b18
Core tap last commit: 2 days ago
Core tap branch: master
HOMEBREW_PREFIX: /usr/local
HOMEBREW_AUTOREMOVE: set
HOMEBREW_CASK_OPTS: []
HOMEBREW_DEVELOPER: set
HOMEBREW_DISPLAY: /private/tmp/com.apple.launchd.D3193WXYad/org.macosforge.xquartz:0
HOMEBREW_MAKE_JOBS: 4
Homebrew Ruby: 2.6.8 => /usr/local/Homebrew/Library/Homebrew/vendor/portable-ruby/2.6.8_1/bin/ruby
CPU: quad-core 64-bit ivybridge
Clang: 12.0.0 build 1200
Git: 2.24.3 => /Applications/Xcode.app/Contents/Developer/usr/bin/git
Curl: 7.64.1 => /usr/bin/curl
macOS: 10.15.7-x86_64
CLT: 12.0.0.32.29
Xcode: 12.4
```

</details>
<details>
<summary>M1 Macbook: Brew Config</summary>

```
HOMEBREW_VERSION: 3.6.10-11-gb683beb
ORIGIN: https://github.com/Homebrew/brew
HEAD: b683bebd24b5003b47e722116eb2a0989e1e7111
Last commit: 2 days ago
Core tap ORIGIN: https://github.com/Homebrew/homebrew-core
Core tap HEAD: 12cb55611f0884a68921be5a0badf29b2b0e3f48
Core tap last commit: 60 minutes ago
Core tap branch: master
HOMEBREW_PREFIX: /opt/homebrew
HOMEBREW_CASK_OPTS: []
HOMEBREW_MAKE_JOBS: 10
Homebrew Ruby: 2.6.10 => /System/Library/Frameworks/Ruby.framework/Versions/2.6/usr/bin/ruby
CPU: 10-core 64-bit arm_firestorm_icestorm
Clang: 14.0.0 build 1400
Git: 2.38.1 => /opt/homebrew/bin/git
Curl: 7.84.0 => /usr/bin/curl
macOS: 13.0-arm64
CLT: 14.1.0.0.1.1666437224
Xcode: N/A
Rosetta 2: false
```

</details>