@MikeMcQuaid 2022-11-14
> I decided to limit this optimization to the core tap because I don't think that any other taps are active enough for this to be a performance issue. Does anyone know if there are any other taps of a similar size out there?

`homebrew-cask`?

> Is this a worthwhile tradeoff to speed up `brew search` and `brew info`?

I think so! Note `hyperfine` is the best tool for doing these sorts of benchmarks.