@Bo98 2022-11-14
> I decided to limit this optimization to the core tap because I don't think that any other taps are active enough for this to be a performance issue. Does anyone know if there are any other taps of a similar size out there?

To this I ask: is this optimisation a (non-trivial) performance penalty for small taps? If not, might as well avoid the question of tap size and do it unconditionally?