@apainintheneck 2022-11-15
> > I decided to limit this optimization to the core tap because I don't think that any other taps are active enough for this to be a performance issue. Does anyone know if there are any other taps of a similar size out there?
> 
> To this I ask: is this optimisation a (non-trivial) performance penalty for small taps? If not, might as well avoid the question of tap size and do it unconditionally?

No, it's trivial. The reason I was avoiding using it for all taps is that it is less thorough. It will miss any formulae that were added less than a month ago and then removed in the meantime; the current algorithm will still catch those. I think that probably doesn't happen that frequently but it's hard to know. It would simplify the code if we decided to do it for everything though.