@apainintheneck 2022-11-15
> * I think you could do this with a single command with e.g. `git diff @{'1 month ago'}` instead of `git log` to get `month_old_commit_hash`.

I didn't know how to do that. Very interesting! It looks like the following should work.

```sh
git diff @{'1 month ago'} --diff-filter=D --name-only -- Formula/{name}.rb
```

> * are you sure we want to use HEAD here rather than e.g. origin/HEAD or origin/master?

I assume the above command uses HEAD by default. I didn't really think too hard about it to be honest. Is there any downside to doing it this way? If we decide to apply this to all taps, would we be able to count on that always being the branch name? Is that enforced anywhere?

> Note `hyperfine` is the best tool for doing these sorts of benchmarks.

I'll keep that in mind for next time.