@apainintheneck 2022-11-15
There's also a part of me that wonders if there will be a gap between `--since='1 month ago'` and `@{'1 month ago'}`. Maybe it'd be better to be explicit? Check for the existence of the file at `@{'30 days ago'}` and then run the git log command if it exists with `--since='31 days ago'`.