@RandomDSdevel 2022-11-15
&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Alternatively or in addition, would using [the cache store](https://github.com/Homebrew/brew/blob/master/Library/Homebrew/cache_store.rb) per-formula help here any?  Of course, that only works between '`brew update`' runs, which invalidate any cached Git logs (unless only new entries were added into the cache incrementally.)  That might well be too many abstractions, though?