@apainintheneck 2022-11-15
I updated the shell command so that now only one is needed as recommended by @MikeMcQuaid. I also changed it so that this optimization applies to all taps as mentioned by @Bo98. I'm still not sure if that's the right idea but the code is simpler which is nice.