@MikeMcQuaid 2022-11-15
> To this I ask: is this optimisation a (non-trivial) performance penalty for small taps? If not, might as well avoid the question of tap size and do it unconditionally?

I'd flip this round: unless you can demonstrate a speed-up on small taps: it's worth just special-casing homebrew-core.

Smaller taps are in general updated less often so otherwise we make this a lot less useful as a feature for the sake of consistency (if there's minimal or no performance benefit).