@MikeMcQuaid 2022-11-15
> I assume the above command uses HEAD by default. I didn't really think too hard about it to be honest. Is there any downside to doing it this way? If we decide to apply this to all taps, would we be able to count on that always being the branch name? Is that enforced anywhere?

You can specify a revision here. I think it's generally best to compare `origin/master` to older commits on `origin/master` to avoid causing bugs when users have weird stuff checked out (which happens not uncommonly).